Implemented BFS.  Given a start and goal hex, it gives you the next step that you should take to get to that hex.  Can be used for pathfinding.
Right now it is unweighted.  It uses the pathnode turtle, and uses the
pathnode's id, which basically means weighting is down by order of
creation.  Adding terrain weighting may be difficult, but I'm not sure we
need it, since it wasn't that huge a factor in the actual battle, and
since we'd be making an unjustified assumption assigning weights to the
terrain.